move structure to separate files
