+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/CoverageTable.java
@@ -1,0 +1,12 @@
+package org.apache.fontbox.ttf.gsub;
+
+abstract class CoverageTable
+{
+    int coverageFormat;
+
+    abstract int getCoverageIndex(int gid);
+
+    abstract int getGlyphId(int index);
+
+    abstract int getSize();
+}
