+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/CoverageTableFormat2.java
@@ -1,0 +1,12 @@
+package org.apache.fontbox.ttf.gsub;
+
+class CoverageTableFormat2 extends CoverageTableFormat1
+{
+    RangeRecord[] rangeRecords;
+
+    @Override
+    public String toString()
+    {
+        return String.format("CoverageTableFormat2[coverageFormat=%d]", coverageFormat);
+    }
+}
