+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/FeatureRecord.java
@@ -1,0 +1,13 @@
+package org.apache.fontbox.ttf.gsub;
+
+class FeatureRecord
+{
+    String featureTag;
+    FeatureTable featureTable;
+
+    @Override
+    public String toString()
+    {
+        return String.format("FeatureRecord[featureTag=%s]", featureTag);
+    }
+}
