+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/FeatureTable.java
@@ -1,0 +1,13 @@
+package org.apache.fontbox.ttf.gsub;
+
+class FeatureTable
+{
+    int[] lookupListIndices;
+
+    @Override
+    public String toString()
+    {
+        return String.format("FeatureTable[lookupListIndiciesCount=%d]",
+                lookupListIndices.length);
+    }
+}
