+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/GlyphSubstitutionTable.java
@@ -19,7 +19,6 @@ package org.apache.fontbox.ttf.gsub;
 
 import java.io.IOException;
 import java.util.ArrayList;
-import java.util.Arrays;
 import java.util.Collection;
 import java.util.Collections;
 import java.util.Comparator;
@@ -679,253 +678,5 @@ public class GlyphSubstitutionTable extends TTFTable
         rangeRecord.endGlyphID = data.readUnsignedShort();
         rangeRecord.startCoverageIndex = data.readUnsignedShort();
         return rangeRecord;
-    }
-
-    static class ScriptRecord
-    {
-        // https://www.microsoft.com/typography/otspec/scripttags.htm
-        String scriptTag;
-        ScriptTable scriptTable;
-
-        @Override
-        public String toString()
-        {
-            return String.format("ScriptRecord[scriptTag=%s]", scriptTag);
-        }
-    }
-
-    static class ScriptTable
-    {
-        LangSysTable defaultLangSysTable;
-        LinkedHashMap<String, LangSysTable> langSysTables;
-
-        @Override
-        public String toString()
-        {
-            return String.format("ScriptTable[hasDefault=%s,langSysRecordsCount=%d]",
-                    defaultLangSysTable != null, langSysTables.size());
-        }
-    }
-
-    static class LangSysRecord
-    {
-        // https://www.microsoft.com/typography/otspec/languagetags.htm
-        String langSysTag;
-        LangSysTable langSysTable;
-
-        @Override
-        public String toString()
-        {
-            return String.format("LangSysRecord[langSysTag=%s]", langSysTag);
-        }
-    }
-
-    static class LangSysTable
-    {
-        int requiredFeatureIndex;
-        int[] featureIndices;
-
-        @Override
-        public String toString()
-        {
-            return String.format("LangSysTable[requiredFeatureIndex=%d]", requiredFeatureIndex);
-        }
-    }
-
-    static class FeatureRecord
-    {
-        String featureTag;
-        FeatureTable featureTable;
-
-        @Override
-        public String toString()
-        {
-            return String.format("FeatureRecord[featureTag=%s]", featureTag);
-        }
-    }
-
-    static class FeatureTable
-    {
-        int[] lookupListIndices;
-
-        @Override
-        public String toString()
-        {
-            return String.format("FeatureTable[lookupListIndiciesCount=%d]",
-                    lookupListIndices.length);
-        }
-    }
-
-    static class LookupTable
-    {
-        int lookupType;
-        int lookupFlag;
-        int markFilteringSet;
-        LookupSubTable[] subTables;
-
-        @Override
-        public String toString()
-        {
-            return String.format("LookupTable[lookupType=%d,lookupFlag=%d,markFilteringSet=%d]",
-                    lookupType, lookupFlag, markFilteringSet);
-        }
-    }
-
-    static abstract class LookupSubTable
-    {
-        int substFormat;
-        CoverageTable coverageTable;
-
-        abstract int doSubstitution(int gid, int coverageIndex);
-    }
-
-    static class LookupTypeSingleSubstFormat1 extends LookupSubTable
-    {
-        short deltaGlyphID;
-
-        @Override
-        int doSubstitution(int gid, int coverageIndex)
-        {
-            return coverageIndex < 0 ? gid : gid + deltaGlyphID;
-        }
-
-        @Override
-        public String toString()
-        {
-            return String.format("LookupTypeSingleSubstFormat1[substFormat=%d,deltaGlyphID=%d]",
-                    substFormat, deltaGlyphID);
-        }
-    }
-
-    static class LookupTypeSingleSubstFormat2 extends LookupSubTable
-    {
-        int[] substituteGlyphIDs;
-
-        @Override
-        int doSubstitution(int gid, int coverageIndex)
-        {
-            return coverageIndex < 0 ? gid : substituteGlyphIDs[coverageIndex];
-        }
-
-        @Override
-        public String toString()
-        {
-            return String.format(
-                    "LookupTypeSingleSubstFormat2[substFormat=%d,substituteGlyphIDs=%s]",
-                    substFormat, Arrays.toString(substituteGlyphIDs));
-        }
-    }
-
-    static class LookupTypeLigatureSubstitutionSubstFormat1 extends LookupSubTable
-    {
-        LigatureSetTable[] ligatureSetTables;
-
-        @Override
-        int doSubstitution(int gid, int coverageIndex)
-        {
-            throw new UnsupportedOperationException();
-        }
-
-        @Override
-        public String toString()
-        {
-            return String.format("%s[substFormat=%d]",
-                    LookupTypeLigatureSubstitutionSubstFormat1.class.getSimpleName(), substFormat);
-        }
-    }
-
-    static abstract class CoverageTable
-    {
-        int coverageFormat;
-
-        abstract int getCoverageIndex(int gid);
-
-        abstract int getGlyphId(int index);
-
-        abstract int getSize();
-    }
-
-    static class CoverageTableFormat1 extends CoverageTable
-    {
-        int[] glyphArray;
-
-        @Override
-        int getCoverageIndex(int gid)
-        {
-            return Arrays.binarySearch(glyphArray, gid);
-        }
-
-        @Override
-        int getGlyphId(int index)
-        {
-            return glyphArray[index];
-        }
-
-        @Override
-        int getSize()
-        {
-            return glyphArray.length;
-        }
-
-        @Override
-        public String toString()
-        {
-            return String.format("CoverageTableFormat1[coverageFormat=%d,glyphArray=%s]",
-                    coverageFormat, Arrays.toString(glyphArray));
-        }
-
-
-    }
-
-    static class CoverageTableFormat2 extends CoverageTableFormat1
-    {
-        RangeRecord[] rangeRecords;
-
-        @Override
-        public String toString()
-        {
-            return String.format("CoverageTableFormat2[coverageFormat=%d]", coverageFormat);
-        }
-    }
-
-    static class RangeRecord
-    {
-        int startGlyphID;
-        int endGlyphID;
-        int startCoverageIndex;
-
-        @Override
-        public String toString()
-        {
-            return String.format("RangeRecord[startGlyphID=%d,endGlyphID=%d,startCoverageIndex=%d]",
-                    startGlyphID, endGlyphID, startCoverageIndex);
-        }
-    }
-
-    static class LigatureSetTable
-    {
-        int ligatureCount;
-        LigatureTable[] ligatureTables;
-
-        @Override
-        public String toString()
-        {
-            return String.format("%s[ligatureCount=%d]", LigatureSetTable.class.getSimpleName(),
-                    ligatureCount);
-        }
-    }
-
-    static class LigatureTable
-    {
-        int ligatureGlyph;
-        int componentCount;
-        int[] componentGlyphIDs;
-
-        @Override
-        public String toString()
-        {
-            return String.format("%s[ligatureGlyph=%d, componentCount=%d]",
-                    LigatureTable.class.getSimpleName(), ligatureGlyph, componentCount);
-        }
     }
 }
