+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/LangSysTable.java
@@ -1,0 +1,13 @@
+package org.apache.fontbox.ttf.gsub;
+
+class LangSysTable
+{
+    int requiredFeatureIndex;
+    int[] featureIndices;
+
+    @Override
+    public String toString()
+    {
+        return String.format("LangSysTable[requiredFeatureIndex=%d]", requiredFeatureIndex);
+    }
+}
