+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/LigatureSetTable.java
@@ -1,0 +1,14 @@
+package org.apache.fontbox.ttf.gsub;
+
+class LigatureSetTable
+{
+    int ligatureCount;
+    LigatureTable[] ligatureTables;
+
+    @Override
+    public String toString()
+    {
+        return String.format("%s[ligatureCount=%d]", LigatureSetTable.class.getSimpleName(),
+                ligatureCount);
+    }
+}
