+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/LookupSubTable.java
@@ -1,0 +1,9 @@
+package org.apache.fontbox.ttf.gsub;
+
+abstract class LookupSubTable
+{
+    int substFormat;
+    CoverageTable coverageTable;
+
+    abstract int doSubstitution(int gid, int coverageIndex);
+}
