+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/LookupTypeLigatureSubstitutionSubstFormat1.java
@@ -1,0 +1,19 @@
+package org.apache.fontbox.ttf.gsub;
+
+class LookupTypeLigatureSubstitutionSubstFormat1 extends LookupSubTable
+{
+    LigatureSetTable[] ligatureSetTables;
+
+    @Override
+    int doSubstitution(int gid, int coverageIndex)
+    {
+        throw new UnsupportedOperationException();
+    }
+
+    @Override
+    public String toString()
+    {
+        return String.format("%s[substFormat=%d]",
+                LookupTypeLigatureSubstitutionSubstFormat1.class.getSimpleName(), substFormat);
+    }
+}
