+++ fontbox/src/main/java/org/apache/fontbox/ttf/gsub/RangeRecord.java
@@ -1,0 +1,15 @@
+package org.apache.fontbox.ttf.gsub;
+
+class RangeRecord
+{
+    int startGlyphID;
+    int endGlyphID;
+    int startCoverageIndex;
+
+    @Override
+    public String toString()
+    {
+        return String.format("RangeRecord[startGlyphID=%d,endGlyphID=%d,startCoverageIndex=%d]",
+                startGlyphID, endGlyphID, startCoverageIndex);
+    }
+}
